refactor(parser): extract inline token dispatcher
